+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 [v3.7.0](https://github.com/sedlatschek/werkbank3/releases/tag/v3.7.0.0)
 - [#2](https://github.com/sedlatschek/werkbank3/issues/2) read VS Code path from registry
 - [#3](https://github.com/sedlatschek/werkbank3/issues/3) set window to foreground on tray icon click
+- [#4](https://github.com/sedlatschek/werkbank3/issues/4) add go environment
 
 [v3.6.0](https://github.com/sedlatschek/werkbank3/releases/tag/v3.6.0.0)
 - fix bug where symlinks could not be deleted